pyproject.toml
[project]
name = "ai-agent-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.5",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.25",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
